Migrate the service-registry to port 8761
This is the default port used by the Eureka when auto-configuring
I thought that there might be a problem when using a non-standard port, but it was
unrelated. The bug encountered (client not registering) was caused by a spurious open
application still listening on the default port (8080), so perhaps it is better to
have the clients (and not just the Eureka server) also on specific ports.